uuid automatizado funcionando
