--- src/main/java/br/com/hbsis/funcionario/EmployeeDTO.java
@@ -1,19 +1,0 @@
-package br.com.hbsis.funcionario;
-
-public class EmployeeDTO {
-
-    private String nome;
-    private EmployeeSavingDTO EmployeeSavingDTO ;
-
-    public EmployeeDTO() {
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-}
