+++ src/main/java/br/com/hbsis/funcionario/FuncionarioService.java
@@ -4,9 +4,18 @@ package br.com.hbsis.funcionario;
 import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.*;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.client.RestTemplate;
 
+import java.util.Arrays;
+import java.util.Objects;
 import java.util.Optional;
+import java.util.UUID;
 
 @Service
 public class FuncionarioService {
@@ -14,23 +23,27 @@ public class FuncionarioService {
     private static final Logger LOGGER = LoggerFactory.getLogger(FuncionarioService.class);
 
     private final IFuncionarioRepository iFuncionarioRepository;
-    private final EmployeeSavingDTO employeeSavingDTO;
 
-    public FuncionarioService(IFuncionarioRepository iFuncionarioRepository, EmployeeSavingDTO employeeSavingDTO) { this.iFuncionarioRepository = iFuncionarioRepository;
-        this.employeeSavingDTO = employeeSavingDTO;
+
+    public FuncionarioService(IFuncionarioRepository iFuncionarioRepository) {
+        this.iFuncionarioRepository = iFuncionarioRepository;
+
     }
 
     public FuncionarioDTO save(FuncionarioDTO funcionarioDTO) {
 
         this.validate(funcionarioDTO);
+        this.createEmployee(funcionarioDTO);
 
         LOGGER.info("Salvando funcionário");
         LOGGER.debug("Funcionário: {}", funcionarioDTO);
 
         Funcionario funcionario = new Funcionario();
-        funcionario.setNomeFuncionario(funcionarioDTO.getNomeFuncionario());
+        funcionario.setNomeFuncionario(funcionarioDTO.getNome());
         funcionario.setEmail(funcionarioDTO.getEmail());
-        funcionario.setUuid(employeeSavingDTO.getEmployeeUuid());
+        funcionario.setUuid(UUID.randomUUID().toString());
+
+
 
         funcionario = this.iFuncionarioRepository.save(funcionario);
 
@@ -49,7 +62,7 @@ public class FuncionarioService {
             throw new IllegalArgumentException("Senha não deve ser nula/vazia");
         }
 
-        if (StringUtils.isEmpty(funcionarioDTO.getNomeFuncionario())) {
+        if (StringUtils.isEmpty(funcionarioDTO.getNome())) {
             throw new IllegalArgumentException("Nome do Funcionario não deve ser nulo/vazio");
         }
     }
@@ -73,7 +86,7 @@ public class FuncionarioService {
             LOGGER.debug("Payload: {}", funcionarioDTO);
             LOGGER.debug("Funcionario Existente: {}", funcionarioExistente);
 
-            funcionarioExistente.setNomeFuncionario(funcionarioDTO.getNomeFuncionario());
+            funcionarioExistente.setNomeFuncionario(funcionarioDTO.getNome());
             funcionarioExistente.setEmail(funcionarioDTO.getEmail());
             funcionarioExistente.setUuid(funcionarioDTO.getUuid());
 
@@ -89,5 +102,16 @@ public class FuncionarioService {
         LOGGER.info("Executando delete para usuário de ID: [{}]", id);
 
         this.iFuncionarioRepository.deleteById(id);
+    }
+
+    private void createEmployee(FuncionarioDTO funcionarioDTO){
+        RestTemplate restTemplate = new RestTemplate();
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("Authorization", "f59ff1b4-1b67-11ea-978f-2e728ce88125");
+        headers.setAccept(Arrays.asList(MediaType.APPLICATION_JSON));
+        HttpEntity entity = new HttpEntity(funcionarioDTO, headers);
+        ResponseEntity<EmployeeSavingDTO> resultadoEmployee = restTemplate.exchange(
+                "http://10.2.54.25:9999/api/employees",  HttpMethod.POST, entity, EmployeeSavingDTO.class);
+        funcionarioDTO.setUuid(Objects.requireNonNull(resultadoEmployee.getBody().getEmployeeUuid()));
     }
 }
